# lab2/Travel_companion_search_service/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/points/create_point": {
        "post": {
          "tags": [
            "points"
          ],
          "summary": "Create New Point",
          "operationId": "create_new_point_points_create_point_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Point"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/points/get_point": {
        "get": {
          "tags": [
            "points"
          ],
          "summary": "Get Point By Name",
          "operationId": "get_point_by_name_points_get_point_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "pointName",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Pointname"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Point"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/points/": {
        "get": {
          "tags": [
            "points"
          ],
          "summary": "Get All Points",
          "operationId": "get_all_points_points__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/routes/create_route": {
        "post": {
          "tags": [
            "routes"
          ],
          "summary": "Create New Route",
          "operationId": "create_new_route_routes_create_route_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/routes/": {
        "get": {
          "tags": [
            "routes"
          ],
          "summary": "Get All Routes",
          "operationId": "get_all_routes_routes__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/routes/get_route": {
        "get": {
          "tags": [
            "routes"
          ],
          "summary": "Get Route By Id",
          "operationId": "get_route_by_id_routes_get_route_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "query",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "Point": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "pointName": {
              "type": "string",
              "title": "Pointname"
            },
            "px": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Px"
            },
            "py": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Py"
            }
          },
          "type": "object",
          "required": [
            "id",
            "pointName",
            "px",
            "py"
          ],
          "title": "Point"
        },
        "PointCreate": {
          "properties": {
            "pointName": {
              "type": "string",
              "title": "Pointname"
            },
            "px": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Px"
            },
            "py": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Py"
            }
          },
          "type": "object",
          "required": [
            "pointName",
            "px",
            "py"
          ],
          "title": "PointCreate"
        },
        "RouteCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "points": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Points"
            },
            "description": {
              "type": "string",
              "title": "Description",
              "default": ""
            }
          },
          "type": "object",
          "required": [
            "name",
            "points"
          ],
          "title": "RouteCreate"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "token"
            }
          }
        }
      }
    }
  }